+++ modules/help.py
@@ -1,34 +1,12 @@
 import discord
 from discord.ext import commands
-from discord import app_commands
 import os
 from dotenv import load_dotenv
 import json
 load_dotenv()
 developerid = os.getenv("developerid")
 
-def has_mod_role():
-    async def predicate(ctx):
-        # Load the setup data from JSON file
-        with open('setup_data.json', 'r') as file:
-            setup_data = json.load(file)
 
-        guild_id = ctx.guild.id
-        setup_info = setup_data.get(str(guild_id))
-
-        if setup_info:
-            mod_role_id = setup_info.get("mod_role_id")
-            if mod_role_id:
-                mod_role = discord.utils.get(ctx.guild.roles, id=mod_role_id)
-                return mod_role is not None and mod_role in ctx.author.roles
-            else:
-                await ctx.send("You do not have the required moderation permissions to run this command!")
-        else:
-            await ctx.send("Pls run d!setup first!")
-
-        return False
-
-    return commands.check(predicate)
 class help(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
