+++ modules/warn.py
@@ -1,38 +1,10 @@
-from discord import app_commands
 from discord.ext import commands
 import discord
-import random
 import asyncio
-import os
 import sqlite3
-from dotenv import load_dotenv
 from datetime import datetime, timedelta
 import dateutil.parser
-import json
-
-
-def has_mod_role():
-    async def predicate(ctx):
-        # Load the setup data from JSON file
-        with open('setup_data.json', 'r') as file:
-            setup_data = json.load(file)
-
-        guild_id = ctx.guild.id
-        setup_info = setup_data.get(str(guild_id))
-
-        if setup_info:
-            mod_role_id = setup_info.get("mod_role_id")
-            if mod_role_id:
-                mod_role = discord.utils.get(ctx.guild.roles, id=mod_role_id)
-                return mod_role is not None and mod_role in ctx.author.roles
-            else:
-                await ctx.send("You do not have the required moderation permissions to run this command!")
-        else:
-            await ctx.send("Pls run d!setup first!")
-
-        return False
-
-    return commands.check(predicate)
+from modules.backend import has_mod_role
 
 
 class warn(commands.Cog):
